+++ inc/pro/functions-display.php
@@ -16,24 +16,16 @@ if ( ! function_exists( 'pagespeed_nav_search' ) ) {
 }
 
 function pagespeed_nav_social_media() {
+    if(!get_theme_mod('show_social_media_link_in_header',false)){
+        return;
+    }
+
 	?>
     <div id="nav-social-media" class="cf nav alignright">
         <ul class="menu">
-			<?php foreach (
-				get_theme_mod( 'social_media_order_nav', array(
-					'facebook',
-					'twitter',
-					'youtube',
-                    'linkedin',
-                    'stack-exchange',
-                    'instagram',
-                    'reddit',
-                    'dribbble',
-                    'github',
-                    'quora',
-                    'google-plus',
-                    'whatsapp'
-				) ) as $link
+			<?php
+            foreach (
+				get_theme_mod( 'social_media_order_nav', array() ) as $link
 			) {
 				echo '<li><a href="' . get_theme_mod( $link, '#' ) . '" class="' . $link . '"><span class="icon he-' . $link . '"></span></a></li>';
 			} ?>